revert a change i made to make the forums HTML compliant.
see bug 2715301

currently, we are doing this:

    <ul>
	<li>thread</li>
	<ul>
	    <li>subthread</li>
	</ul>
    </ul>

which is violates the standard, since bare <ul>'s inside of <ul>'s isn't allowed

we SHOULD do it like this:

    <ul>
      <li>thread
	<ul>
	  <li>subthread</li>
	</ul>
      </li>
    </ul>

but cannot, because we dont know if a thread should have </li> or not here.

technically, we can omit </li> entirely by the HTML standard, but it
violates the XHTML standard, and confuses XHTML DOM parsers, which is rude
(especially for people who use grease monkey to do clever things.

    <ul>
      <li>thread
	<ul>
	  <li>subthread
	</ul>
    </ul>