image file uploading through imgur
if $imgur_client_id is set in config{-local,}.inc, show an "Image
Upload" line on the post form.  when this has a file that php
processed, upload it to imgur and get back a direct image url, and
then set that as the image URL field as if the user put it there
directly.

everything going forward just operates on the image url field as
usual.