log imgur delete url, and force image urls to https
in case a user uploads an image they want to take down, just
removing it from the image url field won't remove it from the
interwebs.  logging the delete url will at least let an admin find
it and give it to the user.

ideally, the imgur url and its delete urls would actually get put
into their own fields in f_messages*, so when the user edits the
post and deletes it, an api call can be made to delete it from imgur
as well (or at least show the delete url to the user).  but this
requires database migrations and some heavy code changes since
images are just converted to html in the message field.

also change image urls to https because imgur supports it and
privacy is good.